# libs/chat-api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login/{providerId}": {
      "get": {
        "operationId": "AuthController_login_v1",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "Configured identity provider ID",
            "schema": {
              "example": "local",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to identity provider"
          },
          "400": {
            "description": "Unsafe callback URL"
          },
          "404": {
            "description": "Unknown provider"
          }
        },
        "summary": "Start OIDC login flow",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/callback/{providerId}": {
      "get": {
        "operationId": "AuthController_callback_v1",
        "parameters": [
          {
            "name": "providerId",
            "required": true,
            "in": "path",
            "description": "Configured identity provider ID",
            "schema": {
              "example": "local",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to app after login"
          },
          "400": {
            "description": "State mismatch or missing tx cookie"
          },
          "502": {
            "description": "Token exchange failed"
          }
        },
        "summary": "OIDC authorization callback",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile"
          },
          "401": {
            "description": "No valid session"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/deployments": {
      "get": {
        "operationId": "DeploymentsController_getDeployments",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of deployment objects from DIAL Core",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DialDeploymentDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "502": {
            "description": "Unexpected response from DIAL Core"
          },
          "503": {
            "description": "DIAL Core is unreachable"
          }
        },
        "summary": "List all available deployments",
        "tags": [
          "deployments"
        ]
      }
    },
    "/api/deployments/{deployment}": {
      "get": {
        "operationId": "DeploymentsController_getDeployment",
        "parameters": [
          {
            "name": "deployment",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DialDeploymentDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Deployment not found"
          },
          "502": {
            "description": "Unexpected response from DIAL Core"
          },
          "503": {
            "description": "DIAL Core is unreachable"
          }
        },
        "summary": "Get a single deployment by name",
        "tags": [
          "deployments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Chat API",
    "description": "REST API for the chat application. Provides endpoints for theme configuration, authentication, and management. All endpoints return appropriate HTTP status codes (200, 400, 401, 403, 404, 502, 503) with descriptive error messages.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "DialDeploymentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "gpt-4o"
          },
          "name": {
            "type": "string",
            "example": "GPT-4o"
          },
          "type": {
            "type": "string",
            "example": "model"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  }
}